# feat(sqlsmith): investigate performance