--- a/src/tests/sqlsmith/README.md
+++ b/src/tests/sqlsmith/README.md
@@ -19,12 +19,15 @@ This test will be run as a unit test:
 
 In the second mode, it will test the entire query handling end-to-end. We provide a CLI tool that represents a Postgres client. You can run this tool via:
 
+Run sqlsmith:
 ```sh
-cargo build # Ensure CLI tool is up to date
-./risedev d # Start cluster
-# Run sqlsmith
+cargo build
+./risedev d
 ./target/debug/sqlsmith test --testdata ./src/tests/sqlsmith/tests/testdata
-# Or run sqlsmith with logs
+```
+
+Run sqlsmith with logs enabled:
+```sh
 RUST_LOG=info ./target/debug/sqlsmith test --testdata ./src/tests/sqlsmith/tests/testdata
 ```