--- a/src/tests/sqlsmith/src/runner.rs
+++ b/src/tests/sqlsmith/src/runner.rs
@@ -25,32 +25,12 @@ pub async fn run(client: &tokio_postgres::Client, testdata: &str, count: usize) 
     let mut rng = rand::rngs::SmallRng::from_entropy();
     let (tables, mviews, setup_sql) = create_tables(&mut rng, testdata, client).await;
 
-    // Test sqlsmith first
     test_sqlsmith(client, &mut rng, tables.clone(), &setup_sql).await;
     tracing::info!("Passed sqlsmith tests");
-
-    // Test batch
-    // Queries we generate are complex, can cause overflow in
-    // local execution mode.
-    client
-        .query("SET query_mode TO distributed;", &[])
-        .await
-        .unwrap();
-    for _ in 0..count {
-        let sql = sql_gen(&mut rng, tables.clone());
-        tracing::info!("Executing: {}", sql);
-        let response = client.query(sql.as_str(), &[]).await;
-        validate_response(&setup_sql, &format!("{};", sql), response);
-    }
-
-    // Test stream
-    for _ in 0..count {
-        let (sql, table) = mview_sql_gen(&mut rng, tables.clone(), "stream_query");
-        tracing::info!("Executing: {}", sql);
-        let response = client.execute(&sql, &[]).await;
-        validate_response(&setup_sql, &format!("{};", sql), response);
-        drop_mview_table(&table, client).await;
-    }
+    test_batch_queries(client, &mut rng, tables.clone(), &setup_sql, count).await;
+    tracing::info!("Passed batch queries");
+    test_stream_queries(client, &mut rng, tables.clone(), &setup_sql, count).await;
+    tracing::info!("Passed stream queries");
 
     drop_tables(&mviews, testdata, client).await;
 }
@@ -64,45 +44,68 @@ pub async fn test_sqlsmith<R: Rng>(
 ) {
     // Test percentage of skipped queries <=5% of sample size.
     let threshold = 0.20; // permit at most 20% of queries to be skipped.
-    let mut batch_skipped = 0;
-    let batch_sample_size = 50;
+    let sample_size = 50;
+
+    let skipped_percentage =
+        test_batch_queries(client, rng, tables.clone(), setup_sql, sample_size).await;
+    if skipped_percentage > threshold {
+        panic!(
+            "percentage of skipped batch queries = {}, threshold: {}",
+            skipped_percentage, threshold
+        );
+    }
+
+    let skipped_percentage =
+        test_stream_queries(client, rng, tables.clone(), setup_sql, sample_size).await;
+    if skipped_percentage > threshold {
+        panic!(
+            "percentage of skipped stream queries = {}, threshold: {}",
+            skipped_percentage, threshold
+        );
+    }
+}
+
+/// Test batch queries, returns skipped query statistics
+/// Runs in distributed mode, since queries can be complex and cause overflow in local execution
+/// mode.
+async fn test_batch_queries<R: Rng>(
+    client: &tokio_postgres::Client,
+    rng: &mut R,
+    tables: Vec<Table>,
+    setup_sql: &str,
+    sample_size: usize,
+) -> f64 {
     client
         .query("SET query_mode TO distributed;", &[])
         .await
         .unwrap();
-    for _ in 0..batch_sample_size {
+    let mut skipped = 0;
+    for _ in 0..sample_size {
         let sql = sql_gen(rng, tables.clone());
         tracing::info!("Executing: {}", sql);
         let response = client.query(sql.as_str(), &[]).await;
-        batch_skipped +=
-            validate_response_with_skip_count(setup_sql, &format!("{};", sql), response);
+        skipped += validate_response(setup_sql, &format!("{};", sql), response);
     }
-    let skipped_percentage = batch_skipped as f64 / batch_sample_size as f64;
-    if skipped_percentage > threshold {
-        panic!(
-            "percentage of skipped batch queries = {}, threshold: {}",
-            skipped_percentage, threshold
-        );
-    }
+    skipped as f64 / sample_size as f64
+}
 
-    let mut stream_skipped = 0;
-    let stream_sample_size = 50;
-    for _ in 0..stream_sample_size {
+/// Test stream queries, returns skipped query statistics
+async fn test_stream_queries<R: Rng>(
+    client: &tokio_postgres::Client,
+    rng: &mut R,
+    tables: Vec<Table>,
+    setup_sql: &str,
+    sample_size: usize,
+) -> f64 {
+    let mut skipped = 0;
+    for _ in 0..sample_size {
         let (sql, table) = mview_sql_gen(rng, tables.clone(), "stream_query");
         tracing::info!("Executing: {}", sql);
         let response = client.execute(&sql, &[]).await;
-        stream_skipped +=
-            validate_response_with_skip_count(setup_sql, &format!("{};", sql), response);
+        skipped += validate_response(setup_sql, &format!("{};", sql), response);
         drop_mview_table(&table, client).await;
     }
-
-    let skipped_percentage = stream_skipped as f64 / stream_sample_size as f64;
-    if skipped_percentage > threshold {
-        panic!(
-            "percentage of skipped batch queries = {}, threshold: {}",
-            skipped_percentage, threshold
-        );
-    }
+    skipped as f64 / sample_size as f64
 }
 
 fn get_seed_table_sql(testdata: &str) -> String {
@@ -172,17 +175,8 @@ async fn drop_tables(mviews: &[Table], testdata: &str, client: &tokio_postgres::
     }
 }
 
-/// Validate client responses
-fn validate_response<_Row>(setup_sql: &str, query: &str, response: Result<_Row, PgError>) {
-    validate_response_with_skip_count(setup_sql, query, response);
-}
-
 /// Validate client responses, returning a count of skipped queries.
-fn validate_response_with_skip_count<_Row>(
-    setup_sql: &str,
-    query: &str,
-    response: Result<_Row, PgError>,
-) -> i64 {
+fn validate_response<_Row>(setup_sql: &str, query: &str, response: Result<_Row, PgError>) -> i64 {
     match response {
         Ok(_) => 0,
         Err(e) => {