I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

Testing locally:
```sh
cargo build
./risedev d
time RUST_LOG=info ./target/debug/sqlsmith test --testdata ./src/tests/sqlsmith/tests/testdata --count 100 | grep "Finished" | sed 's/^.*Fin/Fin/'

```

Seems that bottleneck is in **running the queries**:
```
Finished create table in 5s
Finished generating batch queries in 0s
Finished executing batch queries in 26s
Finished validating batch queries in 0s
Finished generating stream queries in 0s
Finished testing sqlsmith in 52s
Finished generating batch queries in 0s
Finished executing batch queries in 42s
Finished validating batch queries in 0s
Finished testing batch queries in 42s
Finished generating stream queries in 0s
Finished testing stream queries in 51s
Finished cleanup in 0s
RUST_LOG=info ./target/debug/sqlsmith test --testdata  --count 100  0.37s user 0.08s system 0% cpu 2:31.85 total
```

## Checklist

- [ ] I have written necessary rustdoc comments
- [ ] I have added necessary unit tests and integration tests
- [ ] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Documentation

If your pull request contains user-facing changes, please specify the types of the changes, and create a release note. Otherwise, please feel free to remove this section.

### Types of user-facing changes

Please keep the types that apply to your changes, and remove those that do not apply.

* Installation and deployment 
* Connector (sources & sinks)
* SQL commands, functions, and operators
* RisingWave cluster configuration changes
* Other (please specify in the release note below)

### Release note

Please create a release note for your changes. In the release note, focus on the impact on users, and mention the environment or conditions where the impact may occur.

## Refer to a related PR or issue link (optional)