@kwannoel 2022-12-08
Runtime mainly on running queries can't really optimize for now. If sqlsmith takes too long can also reduce recursion depth via `can_recurse`.